Quick Save

+++ src/6thLessonOOP.cpp
@@ -8,7 +8,7 @@ inline uint8_t getMax(uint8_t a, uint8_t b)
 
 int main()
 {
-    uint8_t z = getMax(2, 10);
+    int z = getMax(2, 10);
     cout << "The g = " << z << endl;
     return 0;
 }
